requirements.txt
streamlit>=1.35
pandas>=2.0
gspread>=6.0
google-auth>=2.28
requests>=2.31
